+++ core-api/tests/test_runnables.py
@@ -2,7 +2,7 @@ import pytest
 
 from core_api.build_chains import build_condense_retrieval_chain, build_retrieval_chain, build_summary_chain
 from core_api.dependencies import get_parameterised_retriever, get_tokeniser
-from core_api.runnables import make_chat_prompt_from_messages_runnable
+from redbox.api.runnables import make_chat_prompt_from_messages_runnable
 from redbox.models.chain import ChainInput
 from redbox.models.chat import ChatRoute
 from redbox.models.errors import AIError
